Fix python3.10 problem
