+++ tests/rpcs/test_cfx_rpcs.py
@@ -321,36 +321,6 @@ class TestBlock:
             block = preprocess_block_data(block, use_testnet)
             TypeValidator.validate_typed_dict(block, "BlockData")
 
-class TestPending:
-    @pytest.fixture(scope="class")
-    def future_tx(self, moduled_w3: Web3):
-        nonce = moduled_w3.cfx.get_next_nonce(moduled_w3.cfx.default_account)
-        hash = moduled_w3.cfx.send_transaction({
-            "to": moduled_w3.account.create().address,
-            "value": 100,
-            "nonce": nonce + 1
-        })
-        yield hash
-        moduled_w3.cfx.send_transaction({
-            "to": moduled_w3.account.create().address,
-            "value": 100,
-            "nonce": nonce
-        })
-        hash.executed()
-        
-    
-    def test_get_account_pending_info(self, w3: Web3, address, future_tx):
-        account_pending_info = w3.cfx.get_account_pending_info(address)
-        TypeValidator.validate_typed_dict(account_pending_info, "PendingInfo")
-    
-    def test_get_account_pending_transactions(self, w3: Web3, address, future_tx):
-        nonce = w3.cfx.get_next_nonce(address)
-        info = w3.cfx.get_account_pending_transactions(address, nonce, 1)
-        assert info["firstTxStatus"] == {"pending": "futureNonce"}
-        assert info["pendingCount"] == 1
-        for tx in info["pendingTransactions"]:
-            TypeValidator.validate_typed_dict(tx, "TxData")
-
 def test_check_balance_against_transaction(w3: Web3, address, contract_address):
     payment_info = w3.cfx.check_balance_against_transaction(
         address, contract_address, 10000, 10**9, 0, w3.cfx.epoch_number_by_tag("latest_state")
@@ -360,4 +330,3 @@ def test_check_balance_against_transaction(w3: Web3, address, contract_address):
         address, contract_address, 10000, GDrip(1), 0, w3.cfx.epoch_number_by_tag("latest_state")
     )
     TypeValidator.validate_typed_dict(payment_info, "TransactionPaymentInfo")
-
